Conflicts:
	Makefile
	contrib/bamg/src/Makefile
	contrib/mapx/src/Makefile
	core/src/Makefile
	model/Makefile
	model/finiteelement.cpp
	model/run.sh
	modules/wim/src/Makefile
	nextsim.src

+++ core/include/exporter.hpp
@@ -30,7 +30,7 @@ class Exporter
 {
 public:
 
-	Exporter(std::string const& precision = "double");
+	Exporter(std::string const& precision = "float");
 
     template<typename Type>
     void writeContainer(std::fstream& out, std::vector<Type> const& container);
